requirements.txt
Flask     
opencv-python
Pillow    
numpy         
Werkzeug       
ultralytics